lisp-modules.mgl-pax: include bundled systems
Include the MGL-PAX systems for generating and maintaining
documentation. Add the necessary dependencies of these.

This makes the MGL-PAX package more heavy-weight. Maybe it would be
better to split MGL-PAX up into separate Nix packages in the future.
That might also make it possible to include the test cases without
creating a circular dependency.