lib.systems: accept *-unknown-elf triples
These are the canonical GNU triples for various embedded targets.
Previously, we only accepted the LLVM-style *-none-elf, which was
problematic, because a Matrix user reported needing
riscv64-unknown-elf-gcc for some software, and it makes sense for us
to be able to produce that.  Nixpkgs used to accept such triples, but
it was broken by 3b32c920d56.

With this change, both forms are accepted, and parse equally.
Tested that riscv64-unknown-elf is now parsed correctly, and all
examples still evaluate.

Link: https://github.com/NixOS/nixpkgs/issues/165836
Fixes: 3b32c920d56 ("systems/parse.nix: support eabihf")