nixos/syncthing: Use API to merge / override configurations
If one sets either of `override{Device,folder}s` to false, the jq `*`
operator doesn't merge well the devices and folders, creating duplicate
IDs for folders as observed in #230146. This PR makes the script iterate
via a Bash for loop the devices and folders IDs and merges the keys
using upstream's `curl -X PATCH` support, and support for setting
settings per individual folders and devices via ID in the URL.

Hence this commit fixes #230146.