tree-sitter: Move dynamic libraries for grammar parsers into lib/
The previous version of the code was creating a tree-sitter "grammars"
derivation with all parser's .so file inside the top-level
folder. This is unusual.

Additionally, this prevents other packages from smartly
referencing the parsers. For example, I would expect

  nix-build -E 'with import ./default.nix {};
    pkgs.emacs29.pkgs.withPackages (_:
      [(pkgs.tree-sitter.withPlugins (p: pkgs.tree-sitter.allGrammars))])'

to get an emacs-packages-deps derivation with a lib/ folder and all
parser's .so files there. This would make the

  (add-to-list 'treesit-extra-load-path "$out/lib/")

line in build-support/emacs/wrapper.nix work.