nixos/nftables: Make ruleset.conf be rulesetFile to fix preCheckRuleset
When `rulesetFile` is set, `ruleset.conf` is just a file that includes
that. This makes preCheckRuleset difficult to use since transforming
`ruleset.conf` doesn’t do much anything useful and `rulesetFile` is
probably in the store and so read-only. One could manually `cp
${cfg.rulesetFile} ruleset.conf`, but we might as well do that since, I
think, one wants that in almost all scenarios.

Since the wrapper file when specifying `rulesetFile` is an
implementation detail of this module, I don’t feel this behavior needs
documenting.