nixos/netdata: introduce `deadlineBeforeStopSec`
Previously, we hardcoded a 60 second timer to stop netdata if we didn't have any answer back.
This is wrong and can cause data loss because the SIGTERM sent by systemd can sometimes be not honored.
Which in turn becomes a SIGKILL, causing potential data loss / corruption.

Offer a flag to users and bump the deadline to 2 minutes.