bitwarden: make applying patches easier
Since `sourceRoot` occurs before patching, simply setting `patches` in
the native drv to the same as the npm drv doesn’t work, as paths outside
the `sourceRoot` referenced in the patches don’t exist and cause
applying them to fail. Instead, use `applyPatches` to make `src` have
the patches already applied before `sourceRoot` comes into play. It
would be possible to not use `sourceRoot` and just `cd`, but then the
`Cargo.lock` location needs to be manually specified and it just becomes
more mess than it’s worth. `applyPatches` seems the cleanest way to
achieve this.