nixos-test-driver: use a unique tmpdir for VDE sockets
When running two nixos-test-driver instances that use VLANs on the same
machine, the tmpdir used for vde_switch sockets overlap. This leads to
broken virtual networking in at least one of the two sides, usually with
a zombie vde_switch process.

There are other sources of conflicts/race conditions between
nixos-test-driver instances (e.g. vm-state-* directories) but this one
seems like the easiest to hit by far in my testing.

In theory this shouldn't impact runs on Hydra or via nix-build tests/...
since they should be running in a sandbox with their own tmpdir. In
practice, I'm seeing on Hydra very similar symptoms to the failures I've
observed locally when this race condition is hit, and I'm not completely
sure how to explain it just yet.