houdini: disabled dieWithParent, unsharePid
Due to the nature of the software, it forks and expects children to
outlive the parent, therefore dieWithParent needs to be disabled
Also licensing mechanism seem to be checking processes for it's
licensing helper process running. that process must be same for multiple
instances of houdini itself, therefore Pid namespace has to be the same
for all of them.