nixos/firewall-nftables: unbreak IPv6
For some reason, (at least) ICMPv6 router advertisements fall through
the `ct state vmap { … }` statement that should in theory be
exhaustive, so the input-allow chain which would accept it is never
reached, the message is dropped, and IPv6 is very broken as a result.

Instead of attempting to match on connection tracking state
exhaustively, let the cases to be handled by the input-allow chain fall
through and then unconditionally jump to it directly afterwards. This
seems to work in all cases, including ICMPv6 router advertisements, and
IPv6 works properly again.