thanos: pin to Go 1.19
It panics when running with Go 1.20 with the following error:

```
panic: Something in this program imports go4.org/unsafe/assume-no-moving-gc
to declare that it assumes a non-moving garbage collector,
but your version of go4.org/unsafe/assume-no-moving-gc hasn't been updated
to assert that it's safe against the go1.20 runtime.
If you want to risk it, run with environment variable
ASSUME_NO_MOVING_GC_UNSAFE_RISK_IT_WITH=go1.20 set.
Notably, if go1.20 adds a moving garbage collector,
this program is unsafe to use.
```