gcc: fix cross build with different gcc version
Without the change we pass --disable-bootstrap` and try to build runtime
libraries like `libgomp` or `libgcc` by non-local compiler for cases.

This causes build failures when host compiler does not support target's
required features added in target's gcc version. An example is:

  --host=x86_64 --{build,target}=aarch64:

    gcc-aarch64-unknown-linux-gnu> ../../../gcc-13.1.0/libgcc/soft-fp/brain.h:62:1: error: unable to emulate 'BF'
    gcc-aarch64-unknown-linux-gnu>    62 | typedef float BFtype __attribute__ ((mode (BF)));

The change flips back to more conservative setup: do not disable
bootstrap if it's not a native build. If we want to skip the bootstrap
here we will need to add precise compiler version checks in respective
toolchains.

This fixes an example build of aarch64-gcc-13 with x86_64-gcc-12.