pgadmin4: 7.0 -> 7.1
Upstream has changed their `yarn.lock` file format to v2.
This introduces complications with nixpkgs, which needs the v1
file format. There is also a number of problems with the v2 version
(e.g. reproducibility or missing integrity fields) which makes the
direct use of v2 lockfile impossible.

Due to this change, we will need to convert the lockfile beforehand
and add it to nixpkgs. Due to issues with reproducibility it isn't
feasable to convert the lockfile within the `pgadmin` derivation.

For this commit I added the converted `yarn.lock` file and will add the
corresponding update script, once #231687 has been merged. (This adds
the tool to convert most of the v2 file to v1 and adds metadata to the
lockfile)

Signed-off-by: Florian Brandes <florian.brandes@posteo.de>