betterlockscreen: make dunst optional
When dunst is detected *and* awesomeWM is used (its internal notification system to be exact)
then betterlockscreen waits for dunst to reply on the DBus. Due to dunstctl being in PATH
through the wrapper, dunst will always be tried to reach.

None of the workarounds mentioned in the awesome and betterlockscreen repos do not work with
this version, as dunstctl is always called and has to time out first.

In the current master of BLS there is a change that makes BLS check and env-var first, and
would ignore any dunstctl in PATH.

Once the new release happened this could still be used to reduce the closure size.