Rename redpanda package to redpanda-rpk
The existing package called redpanda only contains the cli tool rpk.

Perhaps redpanda-rpk should be moved out of servers directory since
it is only a tool for interacting with a redpanda server?

I'm unable to build redpanda-server, but I guess it contains what
we need for running redpanda server. Why is that included under the
rpk package passthrough?