docs: Reference packages directly in `allowUnfreePredicate`
This makes the connection from the predicate to the package explicit,
avoiding confusion when `pkg.name != pkgs."${pkg.name}".name`.

The new example demonstrates the subtle difference. To install the
Visual Studio Code package in an FHS-compatible environment, we have to
include nixpkgs' `vscode-fhs` property. But
`pkgs.vscode-fhs.name == "code"`, so `allowUnfreePredicate` ends up
looking like this:

```nix
{
  allowUnfreePredicate = pkg: builtins.elem (lib.getName pkg) [
    "code"
  ];
}
```

By linking to the package explicitly we can instead do this:

```nix
{
  allowUnfreePredicate = pkg: builtins.elem (lib.getName pkg) (map lib.getName [
    vscode-fhs
  ]);
}
```

The result is less confusing in case of package names being different
from the field name in nixpkgs.