nixos/manual: rectify partitioning commands for UEFI/GPT
For GPT partition tables, there is no primary/logical/extended
partition type, and the layout of the mkpart command options slightly
differs depending on whether parted operates on an MBR or a GPT
partition table.

This has so far resulted in the string `primary` being assigned to the
GUID partition name field, which is probably unintended.

Rectify this by giving the GUID partitions sensible names and setting
`fs-type` where applicable.