palemoon: Drop
It has Python 2.x as a build-time dependency, which has been EOL for years and was finally marked as insecure in https://github.com/NixOS/nixpkgs/pull/201859.
As such, it has been unbuildable on master without enabling insecure packages for almost half a year.

Upstream has no plans to migrate their build system away from py2 (though they might accept help with it), so we'll have to drop it:
https://forum.palemoon.org/viewtopic.php?f=62&t=28863&p=232152