util-linux: use real util-linux package on darwin
The util-linux on darwin is an alias for unixtools.util-linux which
contains Apple's variants of the tools. However, the tools are not
compatible with util-linux; for example the command line options differ.

A previous change made util-linux buildable on darwin. This change
removes the alias so util-linux is util-linux everywhere. To access
the Apple tools, use unixtools (e.g unixtools.util-linux).