fiano: init at 1.2.0
Fiano is a suite of Go-based tools for modifying UEFI firmware.

This initial inclusion of fiano omits one package called `fittool`,
since this tool has a separate `modRoot` than the rest and its `go.sum`
is not up to date, so it does not build. I will be working with upstream
to get this fixed.