linux: keep kernel config file after build
We keep the config as sometimes it is useful to check if the kernel
build altered the config. This might happen if the config can't be
used as is in the Nix environment. For example, if rustc is missing,
the CONFIG_RUST=y option would be unset. This helps debugging.