mpv: fix umpv.desktop
mpv ships with a wrapper-script umpv that adds files to the queue of a currently
running mpv instance. This script does not take any options.

The desktop file for it simply substituted mpv with umpv in mpv.desktop,
resuting in an incorrect desktop file with `Exec=umpv --player-operation-mode=pseudo-gui -- %U`.
This resulted in umpv adding bogus files to the queue.