linux: default stdenv.hostPlatform.linux-kernel
With this change, we can do a reasonable default build of Linux for
configurations that do not have a corresponding platforms.nix entry,
and where the user has not explicitly specified the linux-kernel
values.  This allows us to do best effort builds for obscure
architectures (I tested a build for s390x).

The platformName binding has not been used since
70cb7050f5c ("Trying to make the linux kernels also cross-build."),
so I removed it.