XXX wip: linux: enable Rust (rust-for-linux) in kernel builds
This commit enables `pkgs.linuxKernel.manualConfig` to build kernels that have
set CONFIG_RUST=y. For this, it requries rustc and rust-bindgen in path.

To debug problems such as that the kernel build silently removes the
CONFIG_RUST=y option, it is useful to add `make LLVM=1 rustavailable && exit 1`
to the postPatch step and view the result.