takethetime: fetch from PyPI and disable checks
Previously, we built the package from GitHub in order to run the
included tests.
However, because those tests are timing dependent, they could
potentially fail even when the package itself was fine.
Because the tests were not useful, we switched back over to getting the
package from PyPI and only using pythonImportsCheck.