weston: remove unused dependencies
colord support is deprecated and disabled by default.
libGL is redundant with Mesa.
libXcursor is only required when Xwayland is enabled.
libunwind and mtdev is unused since Weston 4.
libxcb and udev are propagated.
pam is unused since Weston 11.

There's no "vaapi" package at all, and there never has been.  This
presumably meant to refer to libva.  This is a very good demonstration
of the problem with the "? null" pattern for optional dependencies!
As a result, VA-API support had been unintentionally disabled by
default this whole time.