nixos/x11/display-managers: don't touch graphical-session.target in xsession-wrapper if the desktop knows how to handle it
This is not correct and will in fact break things because they try to run before the target is reached.
Ideally we'd get rid of it entirely, but WM users rely on this behavior, so allowlist some desktops
to get the sane behavior, and fake the session for the rest until upstreams/NixOS modules catch up.