minimal-bootstrap.live-bootstrap: add helper for fetching useful files
The live-bootstrap project has a handful of files that make building this
software easier during the bootstrapping process. Generally these will be
Makefiles or patches.

This change adds a script to populate hashes in sources.nix and some
helper nix to make it easy to reference the files we want to fetch.