mk-python-derivation.nix: support fixed-point arguments
`buildPythonPackage` and `buildPythonApplication` now accepts
fixed-point arguments, i.e. the `(finalAttrs: { })`style of
input attributes.

The support is provided by `lib.extendMkDerivationModified`.

The value of the `disabled` input argument now assigned to
`passthru.pythonUnsupported`, which determines if the
unsupported error should be thrown.

In the future, we would choose to eliminate such "special" input arguments,
and pass all things directly to `mkDerivation`. Argumnets like this
should go to `passthru`.