gettext: 0.21 -> 0.21.1
Unfortunately, there’s no release announcement for this release. From
the [NEWS][1].

> Version 0.21.1 - October 2022
>
> * Runtime behaviour:
>   - On AIX, locale names with a script or with an uppercase language are now
>     supported.
>     For example, sr_Cyrl_RS.UTF-8 is treated like sr_RS.UTF-8@cyrillic, and
>     EN_US.UTF-8 is treated like en_US.UTF-8.
>
> * The base Unicode standard is now updated to 14.0.0.
>
> * Portability:
>   - Building on macOS 11/arm64 is now supported.
>   - Building on Linux/powerpc64le with glibc ≥ 2.35 is now supported.

The vendored gnulib was also updated, which indirectly fixes a build
failure with clang 16. The version in 0.20 fails to build because it
attempts to assign an incompatible function pointer that should have but
does not have `__attribute__((noreturn))`. On clang 16, this is an error
by default.

[1]: https://git.savannah.gnu.org/gitweb/?p=gettext.git;a=blob_plain;f=NEWS;hb=8b38ee827251cadbb90cb6cb576ae98702566288