updateAutotoolsGnuConfigScriptsHook: preserve original timestamps
Without the change updates to `config.guess` and `config.sub` sometimes
triggers extra rebuilds of artifacts. Ane xample is `libtool` manpages
during the bootstrap on RiscV:

    help2man: can't get `--help' info from libtoolize
    Try `--no-discard-stderr' if option outputs to stderr
    WARNING: 'help2man' is missing on your system.
             You should only need it if you modified a dependency of a man page.
             You may want to install the GNU Help2man package:
             <https://www.gnu.org/software/help2man/>
    make[2]: *** [Makefile:2430: doc/libtoolize.1] Error 127
    make[2]: *** Waiting for unfinished jobs....

Here `libtool` is expected to use prebuilt manpages that come with
release tarball. But as one of it's dependencies was updated by
`updateAutotoolsGnuConfigScriptsHook` hook it now pulls in a missing
extra dependency.

To preserve build system behaviour let's copy timestamps of original
files as is.