boost{175,176,177,178,179,180}: fix build with libc++ 15+
libc++ 15 dropped support for `std::unary_function` and
`std::binary_function` in C++17 and C++20. C++17 is the default for
clang 16, but clang 15 is also affected in those language modes.

This incompatibility is fixed in Boost 1.81, but the late patch for
Boost 1.80 applies to earlier versions and allows those versions to
build with clang 16 (and clang 15 in the affected modes).