nixos/systemd-boot: remove brittle version detection logic
We were relying on the output of `status` which is not very stable.

Instead we rely on systemd to do the version detection logic.
Systemd will not update if version_available <= version_installed

we add --graceful to make sure systemd returns EXIT_SUCCESS in this
scenario.

Unfortunately --graceful also makes it return EXIT_SUCCESS if:
- ESP is not mounted
- Can not edit EFI variables

We alliviate this by introducing a `bootctl is-installed` check.
If that returns true we know the ESP is mounted and that `update`
will not modify any EFI variables.