redmine: Use proper Gemfile and unbreak package
The Gemfile that was added/used in commit 6201512ce2 originates from the
master branch of Redmine instead of its 5.0.5 release tag, which breaks
the Redmine package and makes it unable to start.

For whatever reason the issue wasn't detected during review and also the
tests worked.

Use the Gemfile from the 5.0.5 release tag and do the following changes:
  * Add webrick gem, since it is not included anymore
  * To fix the ffi dependency, add the :mri platform to it
  * Only use the gems for PostgreSQL and MySQL

Also, add a patch that is applied to the Redmine tree to make it
complaint with above changes.

Signed-off-by: Felix Singer <felixsinger@posteo.net>