util-linux: backport bcache checksum patches
In util-linux 2.39, blkid was improved to verify the checksums for
several types of block devices. In the bcache case, this calculation was
incorrect, leading to a regression where valid bcache cache devices
would not be detected any more. In some cases, this can even break the
boot because a required bcache device does not come up any more.

In https://github.com/util-linux/util-linux/pull/2293, a fix was
introduced upstream. Include these patches to make bcache detection work
properly again.