nixos/tests/taskserver: Fix build
The test fails because the way the configuration switch was implemented
back then was by using a dummy configuration and simply activating that
dummy configuration from within the test script.

Nowadays, this doesn't work anymore and fails to typecheck because the
dummy "newServer" will inherit the same value for networking.hostName,
which in turn will generate two attributes for "server":

  > testScriptWithTypes:43: error: Name "server" already defined on line 43
  > [no-redef]
  >     client1: Machine; client2: Machine; server: Machine; server: Machine;

Fortunately, we don't need to do workarounds like this anymore and there
is the "specialisation" option, which allows to do this in a less ugly
way (and it also works with mypy).

Signed-off-by: aszlig <aszlig@nix.build>