nixos/matrix-synapse: add override for PostgreSQL database assertion
This adds an option to override the (useful) assertion that PostgreSQL
must be enabled if Synapse is configured to use a database-connection
to `localhost` (or `::1`, `127.0.0.1`). It was introduced to ensure
that an upgrade to NixOS 20.03 does not result in PostgreSQL being
disabled, as it is no longer automatically enabled in the Synapse
module.

However, this assertion is problematic in more specific use-cases. For
instance, we run multiple Synapse instances on a single machine within
NixOS containers, which share the host's network namespace. In this
setup, a PostgreSQL service is running on localhost, just not defined
as part of the container's configuration. Also, this assertion
prevents running PostgreSQL through means other than the NixOS module,
or using other more complex network configurations.

Thus, this introduces an option to override this check.